#127 Removed class FileHelper.java and moved its test methods
The class FileHelper was removed due to unnecessary externalization. Its only contained method readTextOfPath() is only used in class MailSender.java and therefore should be only accessible by it. The method was moved to the class MailSender.java and the related test methods were moved to the MailSenderTest.java.
