+++ mega-zep-backend/src/test/java/com/gepardec/mega/notification/mail/MailSenderTest.java
@@ -3,6 +3,7 @@ package com.gepardec.mega.notification.mail;
 import io.quarkus.mailer.Mail;
 import io.quarkus.mailer.MockMailbox;
 import io.quarkus.test.junit.QuarkusTest;
+import org.apache.commons.lang3.StringUtils;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -92,5 +93,30 @@ class MailSenderTest {
                 () -> assertEquals(1, sent.size()),
                 () -> assertTrue(sent.get(0).getHtml().startsWith("<p>Hallo " + name)),
                 () -> assertTrue(sent.get(0).getSubject().equals(subjectPrefix.orElse("") + expectedSubject)));
+    }
+
+    @Test
+    void readTextOfPath_correctPath_returnsContent() throws ReflectiveOperationException {
+        assertTrue(StringUtils.isNoneBlank(mailSender.readTextOfPath("notifications/5_plProjectControlling.html")));
+    }
+
+    @Test
+    void readTextOfPath_incorrectPath_returnsNull() {
+        assertNull(mailSender.readTextOfPath("notificatictControlling.html"));
+    }
+
+    @Test
+    void readTextOfPath_nullPath_returnsNull() {
+        assertNull(mailSender.readTextOfPath("null"));
+    }
+
+    @Test
+    void readTextOfPath_BlankPath_returnsNull() {
+        assertNull(mailSender.readTextOfPath(" "));
+    }
+
+    @Test
+    void readTextOfPath_EmptyPath_returnsNull() {
+        assertNull(mailSender.readTextOfPath(""));
     }
 }
